# Improve actor_id for memory efficiency