--- a/pkg/document/change/change.go
+++ b/pkg/document/change/change.go
@@ -84,7 +84,7 @@ func (c *Change) SetServerSeq(serverSeq uint64) {
 }
 
 // SetActor sets the given actorID.
-func (c *Change) SetActor(actor time.ActorID) {
+func (c *Change) SetActor(actor *time.ActorID) {
 	c.id = c.id.SetActor(actor)
 	for _, op := range c.operations {
 		op.SetActor(actor)