--- a/pkg/document/document.go
+++ b/pkg/document/document.go
@@ -154,12 +154,12 @@ func (d *Document) CreateChangePack() *change.Pack {
 
 // SetActor sets actor into this document. This is also applied in the local
 // changes the document has.
-func (d *Document) SetActor(actor time.ActorID) {
+func (d *Document) SetActor(actor *time.ActorID) {
 	d.doc.SetActor(actor)
 }
 
 // ActorID returns ID of the actor currently editing the document.
-func (d *Document) ActorID() time.ActorID {
+func (d *Document) ActorID() *time.ActorID {
 	return d.doc.ActorID()
 }