--- a/pkg/document/operations/edit.go
+++ b/pkg/document/operations/edit.go
@@ -95,7 +95,7 @@ func (e *Edit) ExecutedAt() *time.Ticket {
 }
 
 // SetActor sets the given actor to this operation.
-func (e *Edit) SetActor(actorID time.ActorID) {
+func (e *Edit) SetActor(actorID *time.ActorID) {
 	e.executedAt = e.executedAt.SetActorID(actorID)
 }