--- a/pkg/document/operations/remove.go
+++ b/pkg/document/operations/remove.go
@@ -75,7 +75,7 @@ func (o *Remove) ExecutedAt() *time.Ticket {
 }
 
 // SetActor sets the given actor to this operation.
-func (o *Remove) SetActor(actorID time.ActorID) {
+func (o *Remove) SetActor(actorID *time.ActorID) {
 	o.executedAt = o.executedAt.SetActorID(actorID)
 }