--- a/pkg/document/operations/rich_edit.go
+++ b/pkg/document/operations/rich_edit.go
@@ -102,7 +102,7 @@ func (e *RichEdit) ExecutedAt() *time.Ticket {
 }
 
 // SetActor sets the given actor to this operation.
-func (e *RichEdit) SetActor(actorID time.ActorID) {
+func (e *RichEdit) SetActor(actorID *time.ActorID) {
 	e.executedAt = e.executedAt.SetActorID(actorID)
 }