--- a/pkg/document/operations/select.go
+++ b/pkg/document/operations/select.go
@@ -83,7 +83,7 @@ func (s *Select) ExecutedAt() *time.Ticket {
 }
 
 // SetActor sets the given actor to this operation.
-func (s *Select) SetActor(actorID time.ActorID) {
+func (s *Select) SetActor(actorID *time.ActorID) {
 	s.executedAt = s.executedAt.SetActorID(actorID)
 }