--- a/yorkie/backend/db/mongo/encoder.go
+++ b/yorkie/backend/db/mongo/encoder.go
@@ -27,7 +27,7 @@ import (
 
 func encodeActorID(id time.ActorID) primitive.ObjectID {
 	objectID := primitive.ObjectID{}
-	copy(objectID[:], id[:])
+	copy(objectID[:], id.Bytes())
 	return objectID
 }