--- a/yorkie/backend/db/mongo/encoder.go
+++ b/yorkie/backend/db/mongo/encoder.go
@@ -25,7 +25,7 @@ import (
 	"github.com/yorkie-team/yorkie/yorkie/backend/db"
 )
 
-func encodeActorID(id time.ActorID) primitive.ObjectID {
+func encodeActorID(id *time.ActorID) primitive.ObjectID {
 	objectID := primitive.ObjectID{}
 	copy(objectID[:], id.Bytes())
 	return objectID