--- a/yorkie/backend/sync/memory/pubsub_test.go
+++ b/yorkie/backend/sync/memory/pubsub_test.go
@@ -32,8 +32,12 @@ import (
 )
 
 func TestPubSub(t *testing.T) {
-	actorA := types.Client{ID: time.ActorID{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}}
-	actorB := types.Client{ID: time.ActorID{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1}}
+	idA, err := time.ActorIDFromBytes([]byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0})
+	assert.NoError(t, err)
+	idB, err := time.ActorIDFromBytes([]byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1})
+	assert.NoError(t, err)
+	actorA := types.Client{ID: idA}
+	actorB := types.Client{ID: idB}
 
 	t.Run("publish subscribe test", func(t *testing.T) {
 		pubSub := memory.NewPubSub()