<!--  Thanks for sending a pull request! -->

**What this PR does / why we need it**:
### benchmark result
In particular, there are many improved benchmarks.
```bash
# Improved Percentage
# 25% B/op, 74% allocs/op
Document/text_edit_gc_1000-16                          3         375720418 ns/op        185481637 B/op   2679595 allocs/op
BenchmarkDocument/text_edit_gc_1000-16                 4         295473624 ns/op        137405120 B/op    676571 allocs/op

# 19% B/op, 74% allocs/op
BenchmarkDocument/text_split_gc_1000-16                3         418665897 ns/op        251462381 B/op   2693447 allocs/op
BenchmarkDocument/text_split_gc_1000-16                3         360549476 ns/op        203436376 B/op    691558 allocs/op

# 25% B/op, 87% allocs/op
BenchmarkTextEditing-16                                1        24214185871 ns/op       11357013320 B/op 139867728 allocs/op
BenchmarkTextEditing-16                                1        24855438645 ns/op       8407616768 B/op   16985885 allocs/op
```


```bash
# before
goos: darwin
goarch: amd64
pkg: github.com/yorkie-team/yorkie/test/bench
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDocument/constructor_test-16            1210848               946.8 ns/op           768 B/op         12 allocs/op
BenchmarkDocument/status_test-16                 2275820               525.0 ns/op           736 B/op         10 allocs/op
BenchmarkDocument/equals_test-16                  189476              5726 ns/op            5125 B/op         89 allocs/op
BenchmarkDocument/nested_update_test-16            69469             17124 ns/op           10579 B/op        233 allocs/op
BenchmarkDocument/delete_test-16                   54086             21470 ns/op           12942 B/op        282 allocs/op
BenchmarkDocument/object_test-16                  146668              7450 ns/op            5638 B/op         96 allocs/op
BenchmarkDocument/array_test-16                    43726             27406 ns/op           11492 B/op        262 allocs/op
BenchmarkDocument/text_test-16                     47349             24341 ns/op           12846 B/op        387 allocs/op
BenchmarkDocument/text_composition_test-16         46882             25326 ns/op           16762 B/op        412 allocs/op
BenchmarkDocument/rich_text_test-16                13869             88011 ns/op           38069 B/op       1223 allocs/op
BenchmarkDocument/counter_test-16                  51801             23051 ns/op           12190 B/op        307 allocs/op
BenchmarkDocument/text_edit_gc_100-16                242           4528891 ns/op         2027929 B/op      39098 allocs/op
BenchmarkDocument/text_edit_gc_1000-16                 3         375720418 ns/op        185481637 B/op   2679595 allocs/op
BenchmarkDocument/text_split_gc_100-16               240           4893137 ns/op         2555189 B/op      39654 allocs/op
BenchmarkDocument/text_split_gc_1000-16                3         418665897 ns/op        251462381 B/op   2693447 allocs/op
BenchmarkDocument/text_100-16                       4940            221847 ns/op          111019 B/op       4469 allocs/op
BenchmarkDocument/text_1000-16                       448           2487326 ns/op         1084371 B/op      44073 allocs/op
BenchmarkDocument/array_1000-16                      846           1344225 ns/op         1132270 B/op      12863 allocs/op
BenchmarkDocument/array_10000-16                      91          13701104 ns/op        10224446 B/op     130717 allocs/op
BenchmarkDocument/array_gc_100-16                   7689            147327 ns/op          100028 B/op       1335 allocs/op
BenchmarkDocument/array_gc_1000-16                   727           1602147 ns/op         1199852 B/op      13898 allocs/op
BenchmarkDocument/counter_1000-16                   5088            213115 ns/op          201018 B/op       5522 allocs/op
BenchmarkDocument/counter_10000-16                   537           2149573 ns/op         2203917 B/op      59529 allocs/op
BenchmarkDocument/rich_text_100-16                  4603            239234 ns/op          134858 B/op       4228 allocs/op
BenchmarkDocument/rich_text_1000-16                  471           2478453 ns/op         1308012 B/op      41132 allocs/op
BenchmarkDocument/object_1000-16                     606           1896599 ns/op         1579062 B/op      12930 allocs/op
BenchmarkDocument/object_10000-16                     48          21070885 ns/op        13880597 B/op     131245 allocs/op
BenchmarkSync/memory_sync_10_test-16              133831              9087 ns/op            1319 B/op         38 allocs/op
BenchmarkSync/memory_sync_100_test-16              14070             85402 ns/op            7934 B/op        225 allocs/op
BenchmarkSync/memory_sync_1000_test-16              1437            891660 ns/op           74016 B/op       2051 allocs/op
BenchmarkSync/memory_sync_10000_test-16               84          12756992 ns/op          782391 B/op      20549 allocs/op
BenchmarkSync/etcd_sync_100_test-16             1000000000               0.8759 ns/op          0 B/op          0 allocs/op
BenchmarkTextEditing-16                                1        24214185871 ns/op       11357013320 B/op        139867728 allocs/op
PASS
ok      github.com/yorkie-team/yorkie/test/bench        116.724s

----------------------------------------------------------------------------
# after
goos: darwin
goarch: amd64
pkg: github.com/yorkie-team/yorkie/test/bench
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDocument/constructor_test-16            1000000              1037 ns/op             768 B/op         12 allocs/op
BenchmarkDocument/status_test-16                 2137471               571.7 ns/op           736 B/op         10 allocs/op
BenchmarkDocument/equals_test-16                  188680              6022 ns/op            5101 B/op         88 allocs/op
BenchmarkDocument/nested_update_test-16            75918             15863 ns/op           10435 B/op        227 allocs/op
BenchmarkDocument/delete_test-16                   55570             22514 ns/op           12798 B/op        276 allocs/op
BenchmarkDocument/object_test-16                  144123              8034 ns/op            5590 B/op         94 allocs/op
BenchmarkDocument/array_test-16                    40886             29209 ns/op           11348 B/op        256 allocs/op
BenchmarkDocument/text_test-16                     44160             26109 ns/op           12221 B/op        361 allocs/op
BenchmarkDocument/text_composition_test-16         45868             26063 ns/op           16257 B/op        391 allocs/op
BenchmarkDocument/rich_text_test-16                14240             83165 ns/op           36435 B/op       1155 allocs/op
BenchmarkDocument/counter_test-16                  47032             24481 ns/op           12118 B/op        304 allocs/op
BenchmarkDocument/text_edit_gc_100-16                334           3812379 ns/op         1540549 B/op      18796 allocs/op
BenchmarkDocument/text_edit_gc_1000-16                 4         295473624 ns/op        137405120 B/op    676571 allocs/op
BenchmarkDocument/text_split_gc_100-16               289           4320112 ns/op         2070362 B/op      19451 allocs/op
BenchmarkDocument/text_split_gc_1000-16                3         360549476 ns/op        203436376 B/op    691558 allocs/op
BenchmarkDocument/text_100-16                       4819            227890 ns/op          110996 B/op       4468 allocs/op
BenchmarkDocument/text_1000-16                       486           2427527 ns/op         1084348 B/op      44072 allocs/op
BenchmarkDocument/array_1000-16                      853           1347514 ns/op         1108156 B/op      11862 allocs/op
BenchmarkDocument/array_10000-16                      87          13494112 ns/op         9984544 B/op     120716 allocs/op
BenchmarkDocument/array_gc_100-16                   8364            149093 ns/op           97592 B/op       1234 allocs/op
BenchmarkDocument/array_gc_1000-16                   768           1565314 ns/op         1175692 B/op      12897 allocs/op
BenchmarkDocument/counter_1000-16                   5204            214228 ns/op          200994 B/op       5521 allocs/op
BenchmarkDocument/counter_10000-16                   513           2374633 ns/op         2203892 B/op      59528 allocs/op
BenchmarkDocument/rich_text_100-16                  4316            247958 ns/op          134834 B/op       4227 allocs/op
BenchmarkDocument/rich_text_1000-16                  466           2574912 ns/op         1307989 B/op      41131 allocs/op
BenchmarkDocument/object_1000-16                     628           1907242 ns/op         1554824 B/op      11929 allocs/op
BenchmarkDocument/object_10000-16                     56          20599284 ns/op        13642914 B/op     121252 allocs/op
2022-04-08T23:36:42.700+0900    INFO    default etcd connected, URI: [localhost:2379]
BenchmarkSync/memory_sync_10_test-16              138661              8872 ns/op            1317 B/op         38 allocs/op
BenchmarkSync/memory_sync_100_test-16              14131             84813 ns/op            7936 B/op        225 allocs/op
BenchmarkSync/memory_sync_1000_test-16              1436            879105 ns/op           73777 B/op       2043 allocs/op
BenchmarkSync/memory_sync_10000_test-16               90          12335244 ns/op          782848 B/op      20516 allocs/op
BenchmarkSync/etcd_sync_100_test-16                    6         178307508 ns/op          961506 B/op      16477 allocs/op
BenchmarkTextEditing-16                                1        24855438645 ns/op       8407616768 B/op 16985885 allocs/op
PASS
ok      github.com/yorkie-team/yorkie/test/bench        78.362s
```

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
address #305 

**Special notes for your reviewer**:
[Added MarshalJSON() and UnmarshalJSON() to ActorID struct.](https://github.com/yorkie-team/yorkie/pull/308/files#diff-fd1298d36a83848e81b8421f708a9b715b49ce958182eb7ef06e71dd3030ac0eR127-R147)
ActorID is marshaled somewhere. (An example is attached below).
```go
type Client struct {
	ID           *time.ActorID
	MetadataInfo MetadataInfo
}

func NewClient(encoded []byte) (*Client, error) {
	cli := &Client{}
	err := json.Unmarshal(encoded, cli)  // <---------- here
	if err != nil {
		return nil, err
	}
	return cli, nil
}
```
In this case, bytes are not marshaled because they are private fields. In Go, this can be solved by implementing MarshalJSON and UnmarshalJSON. (https://pkg.go.dev/encoding/json#example-package-CustomMarshalJSON)

However, in this case, whenever the structure field structure is changed, MarshalJSON and UnmarshalJSON must also be changed. With this in mind, we can also consider changing the bytes field to a public field.

What do you think?


**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note

```

**Additional documentation**:

<!--
This section can be blank if this pull request does not require a release note.

Please use the following format for linking documentation:
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```

**Checklist**:
- [x] Added relevant tests or not required
- [x] Didn't break anything